Replacing gulp-traceur with gulp-6to5
Advantages:

* No runtime needed, making the compiled code smaller.

The drawbacks:

* Have to transpile code as 'amd' modules;
* Because of the above, `almond` is now a required production dependency.

Caveats:

`almond` must be placed as the last dependency package declared in the
`bower.json` metadata. Not doing so, any 'amd'-like packages declared below it
will have to be `require`d before they can be used. This is because the
packages are concatenated in that order, making any visible libraries declared
below `almond` to be declared as 'amd' modules. Only 'amd'- or 'umd'-like
libraries are affected by that issue (i.e. Phaser and localForage themselves).

Though, it not being an issue related with these changes, will need to be
sorted out in a near future.