No longer compile scripts with 'compact' option
Seems like some recent changes made to 6to5 (3.3 releases) have this problem.

That's a superfluous feature, as the source code will be 'uglified' on final
builds anyway, so better not make it a project default any longer — should not
be a default from the beginning! This will prevent more of these mysterious
errors spotting in the code later.

Solves issue #1.