Placing Phaser first in `bower.json` manifest
This is getting more confusing.

Phaser uses the 'browserified' version of p2.js physics engine, which makes
more complicated to integrate the framework with building tools like Browserify
or Webpack. Since I don't use (yet) P2 in my games, never tried it thoroughly
to verify this situation. It seems like Almond has some complaints about
Phaser's P2 issues too.

Probably the whole bundling thing and the Babel compilation must be reviewed
too. Removing Almond may not be a good alternative now, though.

Placing Phaser as the first library in the Bower manifest makes it be available
globally to the runtime environment, so referring Phaser with `import` lines in
the sample game is superfluous and, thusly, should be removed. (And those were
incomplete, as some game states missed the library importing.)